backend/preference-mongo-cloudrun/requirements.txt
fastapi
functions-framework
pymongo[srv]
python-dotenv
python-jose[cryptography]
httpx
cachetools
python-multipart
uvicorn
gunicorn
starlette